conditional logic tested and implemnted to tags

+++ test_tags.sh
@@ -22,27 +22,37 @@ NEW_TAG_DATA=$(echo $NEW_TAG | jq -r 'if has("data") then .data else . end')
 NEW_TAG_ID=$(echo $NEW_TAG_DATA | jq '.id')
 echo -e "ID: $NEW_TAG_ID Name: Cool new tag"
 
-# get '/:id' : Get newly made tag by ID
-echo -e "\033[1;4mTesting GET the updated tag:\033[0m"
-curl -s -X GET $BASE_URL/$NEW_TAG_ID | jq
+if [ -n "$NEW_TAG_ID" ]; then
 
-# put /:id' : Update the tag name to Updated tag name
-echo -e "\033[1;4mTesting PUT (update) a tag by ID:\033[0m"
-HTTP_STATUS=$(curl -s -o /dev/null -w "%{http_code}" -X PUT -H "Content-Type: application/json" -d '{"tag_name": "Updated Tag Name"}' $BASE_URL/$NEW_TAG_ID)
-echo "HTTP Status: $HTTP_STATUS"
+    # get '/:id' : Get newly made tag by ID
+    echo -e "\033[1;4mTesting GET the updated tag:\033[0m"
+    curl -s -X GET $BASE_URL/$NEW_TAG_ID | jq
 
-# get '/:id' : Get updated tag by ID
-echo -e "\033[1;4mTesting GET the updated tag:\033[0m"
-curl -s -X GET $BASE_URL/$NEW_TAG_ID | jq
+    # put /:id' : Update the tag name to Updated tag name
+    echo -e "\033[1;4mTesting PUT (update) a tag by ID:\033[0m"
+    HTTP_STATUS=$(curl -s -o /dev/null -w "%{http_code}" -X PUT -H "Content-Type: application/json" -d '{"tag_name": "Updated Tag Name"}' $BASE_URL/$NEW_TAG_ID)
+    echo "HTTP Status: $HTTP_STATUS"
 
-# delete '/:id' : Delete a tag by ID
-echo -e "\033[1;4mTesting DELETE a tag by ID:\033[0m"
-HTTP_STATUS=$(curl -s -o /dev/null -w "%{http_code}" -X DELETE $BASE_URL/$NEW_TAG_ID)
-echo "HTTP Status: $HTTP_STATUS"
+    # get '/:id' : Get updated tag by ID
+    echo -e "\033[1;4mTesting GET the updated tag:\033[0m"
+    curl -s -X GET $BASE_URL/$NEW_TAG_ID | jq
 
-# get '/:id' : expect an error : Get deleted tag by ID 
-echo -e "\033[1;4mTesting GET the deleted tag (expecting an error):\033[0m"
-curl -s -X GET $BASE_URL/$NEW_TAG_ID | jq
+    # delete '/:id' : Delete a tag by ID
+    echo -e "\033[1;4mTesting DELETE a tag by ID:\033[0m"
+    HTTP_STATUS=$(curl -s -o /dev/null -w "%{http_code}" -X DELETE $BASE_URL/$NEW_TAG_ID)
+    echo "HTTP Status: $HTTP_STATUS"
+
+    if is_success $HTTP_STATUS; then
+        # get '/:id' : expect an error : Get deleted tag by ID 
+        echo -e "\033[1;4mTesting GET the deleted tag (expecting an error):\033[0m"
+        curl -s -X GET $BASE_URL/$NEW_TAG_ID | jq
+     else
+        echo "Delete request caused an error"
+    fi
+
+else
+  echo "Post request caused an error"
+fi
 
 # get '/' : GetAll : Verify deletion against all tags
 echo -e "\033[1;4mTesting GET all tags:\033[0m"
